requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
black==22.8.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.3
Django==3.2.10
django-phonenumber-field==6.3.0
django-phonenumbers==1.0.1
djangorestframework==3.11.1
flake8==5.0.4
idna==3.3
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.10.1
phonenumbers==8.12.51
platformdirs==2.5.2
psycopg2==2.8.5
pycodestyle==2.9.1
pyflakes==2.5.0
python-dotenv==0.15.0
pytz==2021.3
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
structlog==20.1.0
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.9
